Reduce `LightPhysics::maximum_distance` to `u8`.
This prevents being able to ask for a huge precalculated light
propagation pattern. There's probably more sanity limit work
to be done here, but this at least makes `fuzz_mesh` not OOM
while building arbitrary blocks.

+++ all-is-cubes-content/src/atrium.rs
@@ -88,7 +88,7 @@ pub(crate) async fn atrium(
         })
         .sky_color(rgb_const!(0.242, 0.617, 0.956) * 1.0)
         .light_physics(all_is_cubes::space::LightPhysics::Rays {
-            maximum_distance: space_bounds.size().to_vector().map(f64::from).length() as u16,
+            maximum_distance: space_bounds.size().to_vector().map(f64::from).length() as u8,
         })
         .build();
